Allow partial simulation
This commit allows a simulation to run from an existing state.

+++ asreview/entry_points/simulate.py
@@ -46,8 +46,12 @@ from asreview.models.query import get_query_model
 from asreview.review.simulate import ReviewSimulate
 from asreview.settings import ASReviewSettings
 from asreview.state.paths import get_data_path
+from asreview.state.paths import get_data_file_path
+from asreview.state.paths import get_feature_matrix_path
 from asreview.state.paths import get_project_file_path
 from asreview.state.utils import init_project_folder_structure
+from asreview.state.errors import StateNotFoundError
+from asreview.state import open_state
 from asreview.types import type_n_queries
 from asreview.utils import get_random_state
 
@@ -66,7 +70,7 @@ ASCII_MSG_SIMULATE = """
 """.format(GITHUB_PAGE, EMAIL_ADDRESS)  # noqa
 
 
-def review_finished(project_path, review_id=None):
+def mark_review_finished(project_path, review_id=None):
     """Mark a review in the project as finished. If no review_id is given,
     mark the first review as finished.
 
@@ -92,6 +96,86 @@ def review_finished(project_path, review_id=None):
 
     with open(get_project_file_path(project_path), 'w') as f:
         json.dump(project_config, f)
+
+
+def _is_review_finished(project_path, review_id=None):
+    """Check if the given review is finished."""
+    project_path = Path(project_path)
+    with open(get_project_file_path(project_path), 'r') as f:
+        project_config = json.load(f)
+
+    if review_id is None:
+        review_index = 0
+    else:
+        review_index = [x['id']
+                        for x in project_config['reviews']].index(review_id)
+
+    review_info = project_config['reviews'][review_index]
+
+    return ('review_finished' in review_info.keys()) & \
+        review_info['review_finished']
+
+
+def _get_dataset_path_from_args(args_dataset):
+    """Remove 'benchmark:' from the dataset name and add .csv suffix.
+
+    Parameters
+    ----------
+    args_dataset : str
+        Name of the dataset.
+
+    Returns
+    -------
+    str
+        Dataset name without 'benchmark:' if it started with that,
+        and with .csv suffix.
+    """
+    if args_dataset.startswith('benchmark:'):
+        args_dataset = args_dataset[10:]
+
+    return Path(args_dataset).with_suffix('.csv').name
+
+
+def _is_partial_simulation(args):
+    """Check if there already is a project file with data of the simulation
+    with given args.
+
+    Arguments
+    ----------
+    args : output of argparse
+
+    Returns
+    -------
+    bool
+        Returns True if there is a project file at args.state_file, with
+        the given dataset, and a review with the given settings, which is not
+        marked as finished.
+    """
+    try:
+        with open_state(args.state_file) as state:
+            settings = state.settings
+    except StateNotFoundError:
+        return False
+
+    # Check if the datasets have the same name.
+    try:
+        project_data_file_name = get_data_file_path(args.state_file).name
+    except Exception:
+        return False
+    args_data_file_name = _get_dataset_path_from_args(args.dataset)
+    if project_data_file_name != args_data_file_name:
+        return False
+
+    # Check if the feature matrix is available.
+    feature_extraction_method = settings.feature_extraction
+    try:
+        feature_matrix_file = get_feature_matrix_path(args.state_file)
+    except FileNotFoundError:
+        return False
+    if not feature_matrix_file.name.startswith(feature_extraction_method):
+        return False
+
+    return True
 
 
 def _set_log_verbosity(verbose):
@@ -130,29 +214,43 @@ class SimulateEntryPoint(BaseEntryPoint):
             raise ValueError("Supply at least one dataset"
                              " with at least one record.")
 
-        # rewrite to ProjectAPI
-        init_project_folder_structure(args.state_file, project_mode='simulate')
+        if not _is_partial_simulation(args):
+            # TODO: refactor these functions into a project module
+            init_project_folder_structure(args.state_file,
+                                          project_mode='simulate')
 
-        # Add the dataset to the project file.
-        as_data.to_csv(
-            Path(
-                get_data_path(args.state_file),
-                Path(args.dataset).with_suffix(".csv").name
+            # Add the dataset to the project file.
+            dataset_path = _get_dataset_path_from_args(args.dataset)
+
+            as_data.to_csv(
+                Path(
+                    get_data_path(args.state_file),
+                    dataset_path
+                )
             )
-        )
+            # Update the project.json.
+            project_file_path = get_project_file_path(args.state_file)
+            with open(project_file_path, 'r') as f:
+                project_json = json.load(f)
+
+            project_json['dataset_path'] = dataset_path
+
+            with open(project_file_path, 'w') as f:
+                json.dump(project_json, f)
 
         # create a new settings object from arguments
-        settings = ASReviewSettings(model=args.model,
-                                    n_instances=args.n_instances,
-                                    n_queries=args.n_queries,
-                                    n_papers=args.n_papers,
-                                    n_prior_included=args.n_prior_included,
-                                    n_prior_excluded=args.n_prior_excluded,
-                                    query_strategy=args.query_strategy,
-                                    balance_strategy=args.balance_strategy,
-                                    feature_extraction=args.feature_extraction,
-                                    mode="simulate",
-                                    data_fp=None)
+        settings = ASReviewSettings(
+            model=args.model,
+            n_instances=args.n_instances,
+            n_queries=args.n_queries,
+            n_papers=args.n_papers,
+            n_prior_included=args.n_prior_included,
+            n_prior_excluded=args.n_prior_excluded,
+            query_strategy=args.query_strategy,
+            balance_strategy=args.balance_strategy,
+            feature_extraction=args.feature_extraction,
+            mode="simulate",
+            data_fp=None)
         settings.from_file(args.config_file)
 
         # Initialize models.
@@ -174,8 +272,8 @@ class SimulateEntryPoint(BaseEntryPoint):
         # LSTM models need embedding matrices.
         if train_model.name.startswith("lstm-"):
             texts = as_data.texts
-            train_model.embedding_matrix = feature_model.get_embedding_matrix(
-                texts, args.embedding_fp)
+            train_model.embedding_matrix = feature_model.\
+                get_embedding_matrix(texts, args.embedding_fp)
 
         # prior knowledge
         if args.prior_idx is not None and args.prior_record_id is not None and \
@@ -188,7 +286,7 @@ class SimulateEntryPoint(BaseEntryPoint):
             prior_idx = convert_id_to_idx(as_data, args.prior_record_id)
 
         print("The following records are prior knowledge:\n")
-        for prior_record_id in args.prior_record_id:
+        for prior_record_id in prior_idx:
             preview = preview_record(as_data.record(prior_record_id))
             print(f"{prior_record_id} - {preview}")
 
@@ -212,7 +310,7 @@ class SimulateEntryPoint(BaseEntryPoint):
         reviewer.review()
 
         # Mark review as finished.
-        review_finished(args.state_file)
+        mark_review_finished(args.state_file)
 
 
 DESCRIPTION_SIMULATE = """
